fix testing issues, can test put, delete

+++ bucketlist_django/bucketlist/views.py
@@ -18,7 +18,7 @@ class BucketListView(APIView):
 
     def post(self, request, format=None):
         serializer = BucketlistSerializer(data=request.data)
-        # create bucketlist if data is valid 
+        # create bucketlist if data is valid
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -39,6 +39,7 @@ class BucketListDetailView(APIView):
         '''returns an instance of a bucketlist object'''
         try:
             return Bucketlist.objects.get(id=id)
+            import pdb; pdb.set_trace()
         except Bucketlist.DoesNotExist:
             raise Http404
 
@@ -46,20 +47,25 @@ class BucketListDetailView(APIView):
         '''returns an individual bucketlist'''
         bucketlist = self.get_object(id)
         serializer = BucketlistSerializer(bucketlist)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, id, format=None):
         '''updates a bucketlist'''
         # get the bucketlist
         bucketlist = self.get_object(id)
-        serializer = BucketlistSerializer(bucketlist, data={'name': request.data['name']}, partial=True)
+        serializer = BucketlistSerializer(bucketlist,
+                                          data={'name': request.data['name']},
+                                          partial=True)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
+            return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id, format=None):
         # get the bucketlist and delete it
         bucketlist = self.get_object(id)
-        bucketlist.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if bucketlist is not None:
+            bucketlist.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_404_NOT_FOUND)
